Keep the keymap loaded for longer
I'm planning to make the server ask the client for commands, rather
than the client sending them unsolicited, in which case the previous
keymap-handling code wouldn't work. The previous code was also buggy
(the options menu, if loaded from the main menu, would load the
keymap even if no game was running); that bug was incidentally fixed
when nhnet_get_commands was modified to work even with no game
running, a feature that's used to implement this new code.