Allow wearing amulets as any monster at all again
This is a revert to vanilla behaviour (which was accidentally broken
in the equipment rewrite). This is mostly so you can use unchanging
into anything.

Rings can be equipped as any monster form, but for monsters without
limbs, they'll need to equip them as some other form and then
polymorph. This is also vanilla behaviour. This patch doesn't change
the behaviour (except for monsters with limbs but no hands; I'm not
sure if there are any of those), but it does improve the messages.