Properly distinguish between game & client options
The behaviour is now much, much, cleaner than before. I don't quite want
to say that it works, yet, but it is closer.

The game and client now keep their own copy of the game options. The
game options are reset when the game initialized and every time
`terminate()` is called, to ensure they go back to defaults for loading
a new game.

The client, when changing options, will change the game's options if a
game is in progress, and its own options otherwise. This means that
options made to one game do not affect future games played on that
client; you must start a new game for this to take place.

Options are now also stored much more frequently: the logfile has a
complete listing of the game's options at the start, and binary saves
all contain a copy of the options.

The character creation settings have been made into birth options.
Future work could move game mode & player name to this model as well.

Some notes:
 1) The binary save is just that. It will break if you so much as
    reorder options in the list.
 2) I've broken the server protocol in two to implement this, since I
    wanted to get what I wanted working with the rest. Fixes to come.
 3) I have not put a lot of thought into connection-only mode. It might
    work, it might not.
 4) `flags.init_role` and co. have now been moved to `u.initrole` and
    co. This removes redundancy at the cost of making statements about
    the lifetime of `u` more difficult to make. I'll let ais523 handle
    this one since he's got a better conception of what things should be
    where in the game side of things.