Generalize file locking
Note: This doesn't compile, because it doesn't convert code that's
about to be deleted anyway to the new syntax.

Note: This probably doesn't work on Windows yet.

The locking code now understands read locks and write locks, in
addition to lock upgrading and downgrading. It also has a new API,
which supports the future implementation of write notification on
files (not yet implemented).