A minimal implementation of the new save system
It's finally here! However, it is /very/ minimal. Although complying
with the save format in the documentation, it does no actual saving
during a game. Thus, loading a game will always load the game as of
T:1, with this implementation. Future changes will refine this to
be able to load at the start of the most recent turn, and then at
the most recent user input. (There's no point in being able to save
with a finer grain than that, obviously.)