Don't update inventory while reloading equipped items
Otherwise, we can get a segfault in xname2(). The item itself seems
fine, but its name is not loaded at that point in the reload sequence.

Ideally, we should be removing all computation from game reload, and
have it just insert appropriate values into the appropriate structs.
That's a hard goal to aim for, though.