Avoid side effects in dorecover()
Most of these changes are to do with the vision system. The problem
is that running it can have side effects on the character's memory,
and /not/ running it gives you a blank screen. So instead we allow
it an exemption from the save stability requirements. This is a
little unsatisfactory.

Also exempted until they can be globals-purged: iflags and toplines.
Neither of these are saved in the save file, so placing save
stability requirements on them just leads to the game not working.