Link common_options statically, rather than dynamically
It's too small for us to worry about the headache of dynamic linking
(and the various headaches that go with it in terms of distributing
DLLs, etc.). Perhaps libnethack_common becomes larger, we can
consider a dynamic link again. The dynamic link infrastructure has
been left in place because of that.