Globals purge: Multiturn actions (including multi itself)
Whew. This was a massive change, but the code feels a lot better for
it. Uses of multi for repeat counts have been moved to the client,
and uses of multi for helplessness have been moved into a set of
helplessness-relted functions. The changes to occupations are even
larger; the command function (doread, etc.) now runs every turn
during an occupation, and occupation state has all been moved to
struct u (apart from flags.occupation, which should probably be in
turnstate but it hasn't been moved yet). Finally, action interruption
(which also used to be implemented using multi) has been moved to
flags.interrupted and flags.incomplete (which should probably also be
in turnstate, but I wanted to get the code working).

Currently broken (by this patch):
- Prayer (marked as unimplemented for the time being)
- Messages for multiturn search, wait