Remove "isstr" parameter from options-handling code
Given that the option parsers are now in a common library, it makes
no sense to have two different APIs for setting options that just
vary in whether the parser is on the client or server.

This also involves removing the options from the database entirely,
which is OK because they're no longer used in normal play. This
probably breaks connection-only mode, but it was probably broken
anyway.