Consistent handling for movement and attack actions
Purges globals: flags.mv, flags.travel, iflags.travel1, flags.run,
and probably more I've forgotten about.

Anything that attacks monsters, or that moves the character sideways
via nonmagical means, now goes through a set of functions that check
the interaction mode. This is a combination of the desires of the
user, and the semantics of the command they give. For instance, if
you apply a pickaxe to a monster, it'll normally ask for confirmation
on a peaceful and attack a hostile outright, but different modes
might produce other actions.

The interaction mode for movement commands can be set by the user
using the "movecommand" option. This also affects other commands that
use the interaction mode, in what is hopefully a DWIM way. For
instance, setting "movecommand" to "pacifist", which will not attack
monsters but will displace them, will make gaze attacks with #monster
not attack monsters either, but autoexplore will remain capable of
automatically displacing tame monsters that are in the way.

"movecommand" replaces and extends the "confirm" and "safe_pet"
options, adding many more possibilities.

This also makes leprechaun dodging work consistently, which is nice.
I added a message to make it clear what's happening.

There are some minor gameplay changes to improve consistency. (Most
notably, steed galloping no longer depends on the occupation state.
This is a strict buff to the player, although a very minor one.)

Currently broken/in need of improvement:
- Attacks are permitted/denied correctly, but the messages could do
  with improvement, and sometimes there's no message at all (e.g.
  trying to push a boulder with item interaction off produces no
  message).
- runmode is broken (it was removed from the server with this patch,
  and hasn't been added back to the client); currently, it's jammed
  at "teleport".
- It'd be nice to be able to document what the individual settings of
  "movecommand" do, somewhere that the user could see them.
- The code for passing the argument to lookaround() works, but is
  hackier than I'd like.