Bugfix for resuming from suspend
This was a reversed test, meaning that the rawsignals setting would
be toggled. No immediate symptoms, but next time you pressed
control-Z, it would be interpreted by the shell rather than by the
game itself (due to rawsignals being off), thus meaning that the game
never got a chance to call doredraw.