Rewrite the message handling code
The previous code was a mess to follow, so I replaced it with some
new code that's easier to understand, and should have fewer bugs.

The only bug I'm currently aware of is if there's a very very long
message (sufficiently large that it overflows the entire message
buffer by itself), followed by another message on the same action.
In this case, the last screenful of the first message will be
scrolled off the screen by the second, without a chance to print a
--More--. This is a pretty minor bug (it hardly ever comes up unless
you specifically try to reproduce it), but worth noting so that it
can be fixed later.

Fixes #80, #87.