Don't crash upon resizing the terminal to smaller than 80x24
We can't produce sensible rendering in these circumstances, but not
crashing means that the user can fix the situation via resizing it
larger again. This implementation works via removing the status
window and sidebar (which are derived windows) altogether, and then
letting the other, non-derived, windows stay in their previous
locations, even if it means their content is drawn offscreen.