Make amnesia lose spells at random.
The implementation is a tad naïve, but will generally result in the same number
of spells lost (the average will be the same, at least). I dunno what the
effects on the variance are but I'm far too lazy to work that out.

Nobody will notice anyway, except one guy that everyone else thinks has variant
paranoia.