Prevent `place_object` from updating vision array on other levels.
This fixes a problem where boulders on a different level were messing with the
vision array.

This is not a full solution to the vision problem generally, because the vision
array on the save-every-turn approach is still very fragile. We are assuming
that it does not change during the loading process. This seems to work, however,
there may be more bugs hiding someplace due to parts of the restore code
breaking this assumption.