Reduce code duplication in menu code
Also improves the code style to avoid things like function-like
macros that assign to their arguments. The major change is the
consolidation of the menu list code in libnethack_common, together
with using it in the client rather than writing out its expansion by
hand all over the place.

This also fixes a memory leak in the extended command menu #?.