Avoid an infinite loop upon hangup save in server play
The server API was incorrectly refusing to allow the client to send
a "force save" command while a game was running. (This was broken
even in NitroHack, where it would refuse a force-save command at a
prompt, leaving the game to time out without being saved; it's worse
in NetHack 4 because it happens outside propts too.) To make things
worse, the server's previous response to an out-of-place command was
to tell the client "restart the connection and send that command
again", which obviously isn't going to be helpful if it just sends
the command again in the same context and turns out unable to run
it again.

Hangup saves now work in network play, although as of this revision,
the client sometimes has trouble sending them.