Fix memory allocation for the crga
The class/race/gender/alignment code was being somewhat inconsistent
with memory use; the C API was leaving the crga allocated a little
longer than the other requests so that the codeflow in the client
happened to work, but the network API code wasn't so accommodating.
This fix makes a copy of the crga rather than relying on it to stay
allocated, and adds in the appropriate xmalloc_cleanup calls in the
client API code so that similar mistakes in the future will fail on
both local and network play, rather than working locally and leaving
the issues in network play undetected.