Remove object symbols from menus
In console play, they really aren't worth the space they take up. In
tiles play, they arguably would be, except that nothing's forcing the
tile size to match the font size, and so it would mess with the
spacing of the menus (quite badly at 32x32, even worse at 48x64).
Simplest is just to remove them altogether and gain an extra couple
of columns.