Move the SDL tiles cursor to a separate texture
It now moves around correctly; there are no copies of the cursor left
behind, because it's erased via blanking the cursor texture, rather
than via overwriting it with an existing (potentially transparent)
tile.

This is possibly less efficient than it should be, but I haven't
noticed any problems in practice.