Index columns starting from 0
The fact that columns are indexed from 1 and rows from 0 is *dumb*.

This probably breaks a bunch of other things in strange and wonderful ways, but
those can be fixed later. We need to bite this bullet at some point. A
surprising amount of code was already assuming 0-indexing so it doesn't need
changing (and just got slightly faster)

I also went and fixed a few indexing errors here and there. The things you
find...