Fix getargdir in grid bug form
This prevents the movement direction being entirely ignored and
re-asked-for whenever moving as a grid bug. This isn't ideal, though;
it works fine for orthogonals, but re-asks for diagonals (and then
refuses to move if diagonals are pressed again). In other words,
everything now works fine given correct input, but the behaviour in
error conditions could use a better UI (as it's a UI issue, that
could be fixed on the client).

Partially fixes #392.