Do a lot of code cleanup.
This fixes a lot of unreachable code, and makes it clear when we're going beyond
standards-compliant code. In general, our code is compliant for X/Open, but in a
few cases in the server code, we go beyond that. In the event someone wants to
port it further, this should make more explicit the cases where we go beyond the
standard (and we definitely don't go far).