Always print messages for items burning in lava.
The previous behaviour didn't make sense because this meant that, at least from
a visible perspective, the timing of object destruction is inconsistent. If you
have life saving, they appear to burn up beforehand; if you do not, they appear
to burn of afterward (this is observable with bones files, which require a
wizard to make because lava is naturally only on special levels, but it could
theoretically happen). Also cases where wizmode cannot observe a behaviour are
dumb.