Ensure we always keep track of all allocated objects
At almost all times, all objects are now on exactly one object
chain (counting the level and individual square object chains as
one chain, because they behave like one). This is achieved using a
"floating object" chain in turnstate for objects that are in transit
between chains.

At the moment, this doesn't do anything useful except check for
memory leaks, but I plan to use it to make object handling exception
safe in a future commit.