Avoid an infinite loop recovering a game that crashes during disclose
This isn't a perfect fix: all writing to a file should attempt to
read from it first, and this doesn't, thus the infinite loop could
still occur given a malicious save file or an engine incompatbility
in which the player dies in one game and not in the other. Still,
it's a one-line fix that doesn't hurt anything and has another
positive side effect (loading a game that's already ended now loads
it at the start of the disclose, rather than attempting to replay the
disclose sequence, meaning you get to see the DYWYPI again, which is
entirely reasonable behaviour for loading an ended game in "play"
rather than "replay" mode).