Rewrite realtime handling
Realtime is now handled in a log-safe way; additionally, the log
contains timestamps for all user input, making it possible to
play back a replay in realtime if we want (or write something like a
log to ttyrec or log to video converter).

Currently, night, moonphase, etc. are updated at the end of each
command, rather than at each user input. The latter is also possible
with this codebase, but leads to some strange timing problems (e.g.
if it changes mid-turn, when do you print the message?)

Moonphase and Friday 13th have't been tested (because it isn't a
weird moon or a Friday 13th today on 6 April 2014).

This also exposed a really insidious bug in memfile.c, fixed in this
patch. I'm glad I decided to use subsecond accuracy, or it might not
have been noticed for another 24 years :-)