BUFSZ purge, part 2: API requests from engine to windowport
The API now uses call-by-name for things like requesting strings from
the user, meaning that the caller gets to choose an appropriate
allocation strategy (e.g. messages when the engine's doing the
calling), and gets to know how much to allocate when doing it. Fixed
size returns that are more complex than an int (such as the return
from "getpos") are now returned by value as structs, rather than by
reference, to avoid confusion about their lifetime.

Known bugs:
- Character creation seems to work, but spouts serious warnings and
  could really do with a rewrite
- Options mostly "work" but could also do with a rewrite; cancelling
  the setting of numerical or string values doesn't work correctly,
  but that's the only known actual bug
- This change make curses_getlin() no longer exception-safe (because
  it's now using allocated memory not a static buffer); this is
  probably easy to fix but that will be a future commit

Bugs that are present, but possibly not the fault of this patch:
- The game over sequence is messed up in terms of allocation,
  particularly the high-score table and "killer" family; segfaults
  are common and can be obtained in mutiple ways
- An extra welcome message displays upon using a count prefix to the
  "drop" command (and probably others too)
- The reminder lines in history that correspond to inventory menus
  could be better-formatted (actually, this was previously broken,
  and this patch breaks it a different way)

Not tested yet:
- JSON client/server code