Fix zeroing-out of objects
In case you can't spot the error in the old code (it took me ages):
the length of the object's name is being measured in units of
sizeof (struct obj), rather than in chars, meaning that the memset
hits random memory somewhere after the end of the object.

This probably explains the random memory corruption I was seeing in
a few specific games. Good thing Grimtooth happened to randomly
generate on D:1 while I was testing under Valgrind, or this might
not have been caught for months. Thanks, Valgrind! Thanks, Grimtooth!