Implement the giant killer rewrite.
The big change here is the elimination of the global variable `killer`. The
killer is now passed as an argument to `done`. Functions in end.c provide help
with formatting a killer message appropriate, and modifiers (like "with the
Amulet" or "while helpless") are added by `done` in a uniform manner.

In most cases, functions like `losehp` where the killer is passed around should
have the formatted killer string passed as an argument, which will then be
passed directly to `done`. The exception is `explode`, although it does not need
to be; the function needs a rewrite anyway and possibly this API could be
corrected when the rewrite is done.

Functions in end.c also provide a way for the game to store delayed killers,
which are now separate based on cause of death (so scrolls of genocide can no
longer turn someone to slime) and handled in a uniform manner. They are stored
on `you` but managed only by end.c as well as the save/restore code.

There are likely huge issues with the killers in specific locations. In
particular, articles are now required, and probably were accidentally ommitted
in a number of places. Any issue with death messages should be corrected.

I would appreciate one or two thorough reviews of this massive patch, due to the
amount of code touched.