Owned pointers should be pointer-to-mutable, not pointer-to-const
Basically, because at least one pointer to the data has to be
mutable, or there is no way to mutate it ever, and freeing is a sort
of mutation (free takes a void *, not a const void *, so the standard
agrees with me on this).  There's no benefit from const'ing the
target of a pointer when you're the owner, either, because the
purpose of consting is to prevent you changing the data behind
someone else's back, and if you're the owner, you're allowed to
change it behind anyone's back whenever you want.