Make death messages more specific, mark user-controlled portions
Unique deaths are hard to measure, and yet quite a tournament
tradition. Let's help the tournament organizers out a bit, as well as
making more deaths distinct ("killed by searching a monster's
interior for a cockatrice corpse", anyone?).

This also contains assorted bugfixes in the code it touches, code
deduplication, and one globals purge ("kickobj" which was used only
for death messages).