If a log has missing/deleted save backups, restore from commands
This is not reliable, of course, but it's better than going into an
infinite loop, which is what the previous code did, and it's useful
in situations where you want to change a save file retroactively
(e.g. to fix a bug in level generation and see the effect on a save
file that already exist, which was the motivating example).