Better death messages for suffocation, explosion, polyself
Apart from deaths to strangulation and suffocation, this is almost
entirely deaths caused via doing something that only kills monsters,
with an amulet of unchanging equipped. The death message is no longer
limited to "stuck in monster form" in such situations.

This contains gameplay changes: in some cases, being hit by a
monster's specific vulnerability would do no damage if you had
unchanging, whereas now, it kills you. (The exception is
self-cancellation as a clay golem, which still has no effect with
unchanging; if you want to die that way, you'll have to find a
gremlin to cancel you instead.) This gameplay change is pretty
minor because it's very rare for players to polyself into golem
forms.