Do not drop light sources when loading new ones.
When reviewing my code, I realized the comment is wrong. We actually *do* need
to correctly load all the light sources in the same order that they are stored
previously, as with the current code, while a save + load is idempotent, light
sources can change from global to local.

As a result, a save + load can change the relative order of a global and a local
light source from the old in-memory version, and then one can change types, so
that a process which does save + load will give a different save from one that
didn't.

I am not sure the best way to fix this. One option would be to make
global light sources store their offsets in the chain. Another would just be to
get rid of the distinction, which my gut says ought to be safe but I am not
sure. It would simplify the save/restore code in a couple of places though.