Use COLNO/ROWNO, not 0, as the sentinel for mx/my.
Previously, the code used mx == 0 to indicate that either a monster was
migrating, or that it was a guard waiting to close its corridor. This is
problematic with the new indexing system.

Instead, switch to COLNO/ROWNO and add checks in various places to avoid
crashes. These values will crash noisily if we use them inappropriately, which I
actually consider a good thing.

I'd prefer not to have sentinel values at all, but to do that properly we really
need to rethink migrations. Guards could be handled with by moving the corridor
information onto the level structure. Both of these are fairly involved, though,
so they're projects for another day.