Fix up the `force` paramater to do what's desired.
The `force` parameter to `water_damage_chain` was actually never `TRUE`, and it
was being used for rust traps' water damage to skip the chance of damage not
being dealt. So it has been removed from `water_damage_chain` and no longer
forces grease to be used up in `water_damage`, which brings uses of that
function closest in line to the previous behaviour of not having the luck-based
saving throw for rust traps.