Clean up the teleport code
(Adapted and expanded from DynaHack commit
5f60575dc7d4227a5a26a52049f27fbef5875ca9.)

Like the DynaHack change, this allows teleports to be cancelled in
the game engine (not yet tested, because the interface disallows
the cancelling). However, I thought I'd spotted a bug in the change,
and thus fixed it. Further investigation shows that it's actually a
vanilla bug (although more exploitable in DynaHack than in vanilla):
leashed pets are teleported to you when you press ^T, even if the
teleport is cancelled or fails due to the Mysterious Force or similar
shenanigans.

This fix also moves the failure message for being anchored by cursed-
leashed pets to after you've selected a destination, to prevent
free curse-testing of applied leashes in zero time via exploiting the
fact that they anchor the player.