Manually merge/squash dtsund/messages-rewrite
This combined patch contains all changes not in
106d9d128fb537b684d4463ac2d0f95dd092ffa8 but in
224e27b18e379d87e3b8aaf1450c6ffe7f52b7b9.

The idea is to rewrite the message system yet again, in the hope that
it works this time; the intermediate versions are not in history
because many of them are badly broken. This change is not perfect by
itself, but should be a starting point for fixing the messages
system.