Implement multiprocess play
This is the infrastructure that will be used for watching live games,
and replaying games while someone else is playing them. It can be
used for other things, too; the only mode of multiprocess play
supported at the moment is to load a game in two processes, both of
which can play it live (sort-of like using tmux). The major benefit
of /that/ is that you can get disconnected and not have to wait to
time out, or you can leave a game running on one connection and
nonetheless play it on another, etc.

Currently broken:
* The server API for server_cancel is not implemented yet;
* Some commands (e.g. inventory) are not server-cancelled properly;
* Sometimes the game freezes for several seconds (likely a locking
  issue)

This has not been tested on tiles (and there are some tiles-specific
codepaths in it, because the way in which you cancel key input is
different for tiles and tty). It has also not been tested on Windows,
but is not expected to work there; the test would be to ensure that
it doesn't break anything in ordinary, single-process play.