Handle zero-time commands in multiprocess play
I was originally going to make them be cancelled out automatically if
another process moved forwards with the game, as happens with time-
consuming commands that are entered simultaneously. However, I like
this way better; the zero-time command freezes time in the game until
it ends, and then the process replays everything that happened since.
This allows you to do things like, say, actually look at the
inventory of someone who is playing very quickly.