Always link against pthreads, even for tty play
We only use one thread for tty play, but always using the thread-safe
version of sigprocmask() is the easiest way to avoid thread-related
UB (libnethack shouldn't care whether it's linked against a thread-
using interface or not). Incidentally, the old code is AFAICT
perfectly fine on Linux, where the restrictions on sigprocmask() in
multithreaded processes are to stop you blocking SIGRTMIN-1 or
SIGRTMIN-2 (signals that aren't available to the user anyway, as can
be seen from the negative numbers, but which do exist and thus are
affected by sigfillset()). However, it's best to avoid UB so as to
keep things working on other systems.

I chose to not use the "-pthread" compiler switch because it
apparently isn't portable between systems. Instead, I emulate its
effect by hand, using "-D_REENTRANT" (the actual "link against
pthreads" part of it is handled by aimake).