Fix another race condition
This one was really subtle (especially as it wasn't in files.c). The
problem is that locking the logfile before recording it in
program_state means that should another process want access to the
logfile in that instant before we record where it is, this process
will dutifully unlock stdin, then wait for the other process to say
that they've got the lock we just "released", forming a deadlock.

There are at least two known race conditions left:
- There's no lock preventing the file lengthening between
  log_replay_input and log_reply_no_more_options, and if it happens
  to lengthen at that precise moment, we get a desync;
- SDL_PushEvent isn't as async-signal-safe as I thought it was; it
  will freeze the process if the signal happens to arrive inside a
  call to malloc(). This is frustrating and possibly hard to fix,
  given that merging everything into one event loop is the normal
  way to make a function async-signal-safe. If SDL doesn't provide
  a method of doing that, we have to generate fake input by bouncing
  it off the OS kernel somehow (this was the trick that was commonly
  used before pselect() was invented for the POSIXy equivalent).