Fixes to getch() interruption in the SDL uncursed backend
This fixes the race condition on malloc, and also changes the way
select() is used because the old version couldn't possibly work on
Windows due to API differences (I'm not sure whether the new version
does or not, but it has a better chance). This is done by replacing
the "correct" threading/event loop style with polling, which upsets
me; however, SDL has no async-signal-safe way to post events (because
it uses malloc for events, and malloc is not async-signal-safe), so
polling is needed somewhere, and SDL does polling every 10ms anyway.
Thus, we simply use the same poll interval as SDL, and get only a
minimal performance penalty.