Async-signal-safety in multiprocess play
How is sigtimedwait() not async-signal-safe? This replaces it with
pselect, and does a few optimizations along the way.

This is still not 100% correct because there's no sigprocmask variant
that's both thread-safe and async-signal-safe. I'm hoping that this
is a mistake in the documentation for pthread_sigmask on my system
(i.e. it is safe, but isn't marked as such); it says it's identical
to sigprocmask, which is safe, after all.

There's still a race condition lurking here; I noticed this while
debugging that. (The problem is that it's possible for three
processes to set up a never-ending cycle of 1-then-2 signals. I'm
unclear whether it's better to prevent such chains starting, or to
prevent them lasting when they start; I'm also unclear on how it
managed to get started in the first place.)