Remove all known UB from the multiprocess play code
I had to delve into the dark corners of POSIX to do this. As far as I
can tell, the only portable way to unblock just the signals that are
being handled, while in a signal handler in a multithreaded process,
is to use the getcontext() buffer passed in a sigaction callback
(which is basically a sigsetjmp buffer, except that the structure is
not 100% opaque; the signal mask and stack are the only portably
accessable fields). getcontext() was removed from POSIX because
setcontext() doesn't actually work, but the third argument to
sigaction() is still there, even if this is the only portable use to
be made of it.

The other callbacks (apart from the SIGALRM callback) were converted
over to the three-argument form too, because I'm planning to use the
second argument in a future commit.

This change also ensures we don't overwrite other people's errnos.
Quite a few errno-setting calls are defined behaviour inside a signal
handler, so setting and reading errno must be too.