Let's try something else with respect to multiprocess play
The previous method didn't scale above more than 4 or 5 processes in
tty play; there were just too many pointless retransmissions. This
alternative version is immune to that, and also avoids a bunch of
problems that the other method had (some of which had workarounds,
some of which didn't).

Sadly, though, it still isn't working properly for 3 or more
processes, and there doesn't seem to be any reason for this. I even
traced the syscalls, and they show something impossible happening
(pselect() blocking a signal that isn't in its signal mask), and
something that I can't reproduce outside the NH4 setting. I suspect
I'm missing something. Nonetheless, commiting this for now.