Don't get stuck forever if something goes wrong in multiprocess play
This prevents the alarm getting clobbered by recursive calls.

I've tested this implementation, and it seems pretty reliable
(although not 100% reliable; connecting to a game can be a little
hairy). I managed to play the same game from 30 processes at once,
with a load average of only around 4 on my netbook.

The main brokennesses in multiprocess play at the moment are the
game over code (if one process quits, the others can keep right on
playing, because although the game is marked as over, the other
processes do not realise this), and messages (the behaviour of
--More-- is all over the place).