Make libuncursed a bit more robust against implausible input
Things like hanging up a terminal while it's in the process of being
resized cause problems; also, various obviously incorrect operations
now have the effect of "return ERR" rather than "corrupt memory".
This also fixes an off-by-one in a plausibility check that already
existed in the code, and turns on stack protectors (to help find what
was causing the memory corruption).