Rewrite window layout code and UI flags
The game now handles very large and very small windows more sensibly.
Very large windows have a fourth window, extrawin, below the status
area; this isn't used for anything yet, but allows the inventory to
expand to the full terminal height. When the sidebar isn't drawn
(either because the terminal width is something awkward like 84, or
because the user turned it off), the map will be horizontally
centered. In tiles mode, the map will expand to take up more space
according to how large the tiles are.

Very small terminals still don't render correctly, but will cause
fewer crashes (for instance, the game no longer crashes when opening
inventory in a 2x2 window). Sub-80x24 terminals now produce better
feedback upon game load (the game refuses to load until the terminal
is sized larger, allows the user to size it larger, and allows the
program to be quit if the terminal can't be sized larger for some
reason).

This also contains several fixes to the rendering library, and
(because they were being rewritten anyway) to UI flags; several
dead options have been removed. (Dead options are more dangerous
than dead globals, in a way, because players may try to change
them on the assumption that they do something.)

Fixes #249, #380.