Default bindings for "quiver": {Q or #quiver} rather than {Q}
The main reason for this is to create ambiguity with #quit, making
it rather harder to quit by mistake (you need to type all four
letters of "quit", or else use the S!y method); this should help
out vanilla players who are in the habit of typing #q. This is a
default binding change, so players who really want to accidentally
quit their games can change it back by unbinding #quiver from
the quiver command.