Bugfixes for force-attacking terrain
There are two main fixes here: one is that we don't call
action_completed if the force-attack is interpreted as an occupation
(e.g. attacking a wall with a pick-axe is interpreted as an attempt
to dig, which requires an occuption to be set), and the other is that
trees and axes are now handled correctly (both the success codepath -
using an axe on a tree - and the failure codepaths involving other
combinations of items).

Currently still broken/dubious: using a pick-axe on an upwards
ladder, which is interpreted as an attempt to dig that misses. It's
unclear what the behaviour here should be; there are several options,
including leaving it the same.