Implement mouse support in libuncursed
This isn't used by NetHack yet, but is used by test_uncursed. Not
tested with tiles; but it is tested with SDL and gnome-terminal,
lxterminal, urxvt, xterm, tmux, screen.

It doesn't fully work in Konsole, for two reasons:
- Konsole only reports mouse motion while some button is held down
  (and thus "hover" events are reported only as a result of "drag
  into" events;
- Konsole appears to have a mistake in its coordinate calculations,
  meaning that clicks are displaced by (on my computer) something
  like five or six pixels.

The first issue would be partially possible to work around via
detecting that hover was broken if two clicks were detected in
different locations with no mouse movement in between; the second
issue, however, is beyond libuncursed's control.  Konsole users are
recommended to turn off mouse support in their applications.

Not tested with PuTTY, because I don't have a network connection
handy right now.

Huh, it's convenient/interesting that /all/ the terminal emulators
we support have mouse support. (The Linux console doesn't support
the mouse-support codes we use, meaning we get no features rather
than a broken subset; this is probably for the best.)