When mouse support is turned on, place close buttons on window borders
I considered allowing windows to be closed via clicking outside them
(which would require changes to libuncursed to be done effectively),
but decided against it because of the risk of cancelling a wish that
way, or similarly major issues. A 1x1 target is unlikely to be
clicked on by accident.