Allow farlook via mouse hover
This actually requires quite the new framework to work. The extra
window now has a use - showing control hints and farlook results -
and the map now responds to clicks (currently with "Unknown command
User256" or the like because the clicks aren't coded to do anything
yet). Currently, extrawin's only used for farlook and for getpos's
help text, but I hope to expand it to many other contexts.