Move save/quit interface logic from the server to the client
There were bits of it floating around in both places, and the effect
was that I couldn't get the menus working properly without a bunch of
duplicated code. So let's just go simplify everything by getting the
client to handle the user interface and the server to handle the
effects on the game, which is how things should be.