Build system improvements
This upgrades to aimake version 3.3, and adapts the build system to
take advantage of its new features (one of which is "features",
meaning that the tiles and server executables are now optional, and
that users can switch between the provided libjansson and their own
copy of it).

This also updates the README and FAQ, because the build system
improvements meant they needed substantial changes anyway.

Fixes #478.