Fix #229, #230
A new "animation" option controls the relation between the passage of
time in-game and out-of-game. "instant" is the old behaviour. With
the other choices, the game will pause for a short time (or a long
time, if you chose that option) every turn within an occupation
callback. (TODO: Perhaps it should do that for helplessness too.)
The action can be interrupted with a key, or a click on the map.
(And no matter the setting, it can be interrupted using a prefix
count.)