Fix #403 (cursor flicker issues, including jumping on tiles)
This was a combination of a bunch of different things: a bug in
libuncursed, and incredibly fast and loose messing around with the
cursor in nhcurses. The basic problem is that turning the cursor on
scrolls any tiles window that might overlap the cursor location, which
is normally desired behaviour. The libuncursed bug was that it also
scrolled when turning the cursor off, too; but that isn't enough to
reproduce the bug on its own (it just makes it more visible). Now, the
code in nhcurses does a bunch of "remember the cursor state, turn it
off, show a window, hide the window, turn the cursor back to its
previous state"; notably, it turns the cursor back on without placing
it anywhere, and it may well be over the tiles region at the time.

This commit fixes it by not actually showing the cursor until the
last possible moment, when (presumably) it's in the right place.